backend/requirements.txt
# Django
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.7.0
django-filter==25.1

# Autenticação JWT
djangorestframework-simplejwt==5.5.0

# Banco de dados
psycopg2-binary==2.9.10

# Variáveis de ambiente
python-decouple==3.8

# Servidor de produção
gunicorn==23.0.0

# QR Code (futuro M7)
# qrcode==8.0

# PDF (futuro M7)
# reportlab==4.4.0

# Utilitários
Pillow==11.2.1
